refactor: CouponProperties 삭제
- 중간 피드백 전달에서 지적된 CouponProperties를 삭제하고 쿠폰 할인 validation 설정값을 예전대로 Enum에 static final로 선언함.
- 파일 삭제에 따른 static filed import 수정.
- 기타 공백 수정 및 빈 import 삭제
